Add uniform noise

Adds uniform noise to values, generating replicates with noise added to the original.

Usage

add_uniform_noise(value, n_rep, sd, use_zero = FALSE)

Arguments

value

a single or vector of numeric values

n_rep

the number of replicates to make (numeric). Default is 1.

sd

the standard deviation of the data

use_zero

logical, should returned values be around zero or not?

Value

numeric matrix