information-content-informed kendall tau

Given a data-matrix, computes the information-theoretic Kendall-tau-b between all samples.

ici_kendalltau(
  data_matrix,
  exclude_na = TRUE,
  exclude_inf = TRUE,
  exclude_0 = TRUE,
  zero_value = 0,
  perspective = "global",
  scale_max = TRUE,
  diag_good = TRUE,
  check_timing = FALSE
)

Arguments

data_matrix
samples are rows, features are columns

exclude_na
should NA values be treated as NA?

exclude_inf
should Inf values be treated as NA?

exclude_0
should zero values be treated as NA?

zero_value
what is the actual zero value?

perspective
how to treat missing data in denominator and ties, see details

scale_max
should everything be scaled compared to the maximum correlation?

diag_not_na
should the diagonal entries reflect how many entries in the sample were "good"?

check_timings
should we try to estimate run time for full dataset? (default is FALSE)

Value

numeric

Details

For more details, see the ICI-Kendall-tau vignette: vignette("ici-kendalltau", package = "visualizationQualityControl")

When check_timings = TRUE, 5 random pairwise comparisons will be run to generate timings on a single core, and then estimates of how long the full set will take are calculated. The data is returned as a data.frame, and will be on the low side, but it should provide you with a good idea of how long your data will take.